+++ db/migrate/20170908135101_create_quarter_months.rb
@@ -1,15 +1,16 @@
-class CreateQuarters < ActiveRecord::Migration[5.1]
+class CreateQuarterMonths < ActiveRecord::Migration[5.1]
   def change
-    create_table :quarters do |t|
+    create_table :quarter_months do |t|
+      t.integer :quarter_id
       t.integer :year, null: false
-      t.integer :quarter_number, null: false
-      t.bigint :bigint, null: false, default: 0
+      t.integer :month_number, null: false
       t.bigint :regular_files, null: false, default: 0
       t.bigint :regular_file_size, null: false, default: 0
       t.bigint :regular_total_files, null: false, default: 0
       t.bigint :regular_total_file_size, null: false, default: 0
+      t.index [:year, :month_number], unique: true
+      t.index :quarter_id
       t.timestamps
-      t.index [:year, :quarter_number], unique: true
     end
   end
 end