+++ db/migrate/20170908175328_add_date_index_to_dataset_file_audits.rb
@@ -1,0 +1,5 @@
+class AddDateIndexToDatasetFileAudits < ActiveRecord::Migration[5.1]
+  def change
+    add_index :dataset_file_audits, :created_at
+  end
+end